refactor sidebar (component and model)
