+++ frontend-project-template/{{cookiecutter.project_slug}}/src/frontend/app.py
@@ -1,17 +1,18 @@
+from backend.sum import sum_values
 import streamlit as st
 from common.logger import setup_logger
-from frontend.components.side_bar import my_func
-from frontend.layout.page_config import set_page_config
+from frontend.components.sidebar import SideBar
+from frontend.misc.page_config import set_page_config
 
 
 logger = setup_logger()
 set_page_config()
 st.title("FrontEnd frontend_app template: Work in progress...")
 
-side_bar = st.sidebar
-first_number = side_bar.number_input("Introduce un numero")
-second_number = side_bar.number_input("Introduce otro numero")
+side_bar = SideBar()
+side_bar.render()
 
 if st.button("Calculate sum"):
-    sum_result = my_func(first_number, second_number)
+    side_bar.validate_model()
+    sum_result = sum_values(side_bar.model)
     st.write(f"Sum of values: {sum_result}")
