--- frontend-project-template/{{cookiecutter.project_slug}}/src/frontend/components/side_bar.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-
-def my_func(a: int, b: int) -> int:
-    return a + b
