+++ frontend-project-template/{{cookiecutter.project_slug}}/src/frontend/misc/page_config.py
