+++ frontend-project-template/{{cookiecutter.project_slug}}/src/frontend/models/sidebar.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class SideBarModel(BaseModel):
+    first_number: float
+    second_number: float
